requirements.txt
# requirements.txt
langchain==0.3.2
langgraph==0.2.34
openai==1.51.2
python-dotenv==1.0.1
langchain-openai=0.2.10